// server/modules/Enums/position.ts
export enum Position {
  ASSOCIATE = 'Associate',
  ANALYST = 'Analyst',
  SENIOR_ANALYST = 'Senior Analyst',
  MANAGER = 'Manager',
  SENIOR_MANAGER = 'Senior Manager',
  DIRECTOR = 'Director',
  VICE_PRESIDENT = 'Vice President',
  SENIOR_VICE_PRESIDENT = 'Senior Vice President',
  PRESIDENT = 'President',
}